restructure properties
